PP-4650 Move applepay package into new wallets package.
- Move `applepay` package into new `wallets` package to aid organisation and
share common code within the `wallets` package.
- Renamed `PaymentInfo` to `WalletPaymentInfo` and moved into `wallets/model`.
- Moved end-point for applepay to `/wallets/apple`.
