+++ src/main/java/uk/gov/pay/connector/wallets/WalletType.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.applepay;
+package uk.gov.pay.connector.wallets;
 
 public enum WalletType {
     APPLE_PAY
