+++ src/main/java/uk/gov/pay/connector/wallets/applepay/InvalidKeyException.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.applepay;
+package uk.gov.pay.connector.wallets.applepay;
 
 import javax.ws.rs.BadRequestException;
