+++ src/test/java/uk/gov/pay/connector/model/domain/ChargeEntityFixture.java
@@ -2,7 +2,7 @@ package uk.gov.pay.connector.model.domain;
 
 import com.google.common.collect.ImmutableMap;
 import uk.gov.pay.commons.model.SupportedLanguage;
-import uk.gov.pay.connector.applepay.WalletType;
+import uk.gov.pay.connector.wallets.WalletType;
 import uk.gov.pay.connector.charge.model.ServicePaymentReference;
 import uk.gov.pay.connector.charge.model.domain.Auth3dsDetailsEntity;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
