+++ src/test/java/uk/gov/pay/connector/model/domain/applepay/ApplePayDecryptedPaymentDataFixture.java
@@ -1,13 +1,13 @@
 package uk.gov.pay.connector.model.domain.applepay;
 
-import uk.gov.pay.connector.applepay.AppleDecryptedPaymentData;
-import uk.gov.pay.connector.common.model.api.PaymentInfo;
 import uk.gov.pay.connector.gateway.model.PayersCardType;
+import uk.gov.pay.connector.wallets.applepay.AppleDecryptedPaymentData;
+import uk.gov.pay.connector.wallets.model.WalletPaymentInfo;
 
 import java.time.LocalDate;
 
 public final class ApplePayDecryptedPaymentDataFixture {
-    private PaymentInfo applePaymentInfo = new PaymentInfo(
+    private WalletPaymentInfo applePaymentInfo = new WalletPaymentInfo(
             "4242",
             "visa",
             PayersCardType.DEBIT,
@@ -30,11 +30,11 @@ public final class ApplePayDecryptedPaymentDataFixture {
         return new ApplePayDecryptedPaymentDataFixture();
     }
 
-    public PaymentInfo getApplePaymentInfo() {
+    public WalletPaymentInfo getApplePaymentInfo() {
         return applePaymentInfo;
     }
 
-    public ApplePayDecryptedPaymentDataFixture withApplePaymentInfo(PaymentInfo applePaymentInfo) {
+    public ApplePayDecryptedPaymentDataFixture withApplePaymentInfo(WalletPaymentInfo applePaymentInfo) {
         this.applePaymentInfo = applePaymentInfo;
         return this;
     }
